+++ AssemblyInfo.cs
@@ -1,10 +1,10 @@
 using System.Reflection;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Coco 桌宠")]
-[assembly: AssemblyDescription("可拖动、可互动的 Windows Coco 桌面宠物")]
+[assembly: AssemblyTitle("Coco Desktop Pet")]
+[assembly: AssemblyDescription("A draggable and interactive Coco desktop pet for Windows")]
 [assembly: AssemblyCompany("Coco")]
-[assembly: AssemblyProduct("Coco 桌宠")]
+[assembly: AssemblyProduct("Coco Desktop Pet")]
 [assembly: AssemblyCopyright("Copyright © 2026")]
 [assembly: ComVisible(false)]
 [assembly: Guid("bfc063b2-1a76-4862-8fc5-251ab3e24170")]